+++ tests/Rule/Performance/DisabledCallsInLoopsRule/Fixtures/ArrayMergeRecursive.php
@@ -6,7 +6,7 @@ namespace Efabrica\PHPStanRules\Tests\Rule\Performance\DisabledCallsInLoopsRule\
 
 final class ArrayMergeRecursive
 {
-    public function callArrayMergeInFor(array $data): array
+    public function callArrayMergeRecursiveInFor(array $data): array
     {
         $result = [];
         for ($i = 0; $i < 100; $i++) {
@@ -15,7 +15,7 @@ final class ArrayMergeRecursive
         return $result;
     }
 
-    public function callArrayMergeInForeach(array $data): array
+    public function callArrayMergeRecursiveInForeach(array $data): array
     {
         $result = [];
         foreach ($data as $row) {
@@ -24,7 +24,29 @@ final class ArrayMergeRecursive
         return $result;
     }
 
-    public function callArrayMergeOnce(array $data): array
+    public function callArrayMergeRecursiveInWhile(array $data): array
+    {
+        $result = [];
+        $i = 0;
+        while (isset($data[$i])) {
+            $result = array_merge_recursive($result, $data[$i]);
+            $i++;
+        }
+        return $result;
+    }
+
+    public function callArrayMergeRecursiveInDoWhile(array $data): array
+    {
+        $result = [];
+        $i = 0;
+        do {
+            $result = array_merge_recursive($result, $data[$i]);
+            $i++;
+        } while (isset($data[$i]));
+        return $result;
+    }
+
+    public function callArrayMergeRecursiveOnce(array $data): array
     {
         return array_merge_recursive([], ...$data);
     }